Refactor: Move application-specific runtime beans to system module (#14)
