+++ backend-system/src/main/java/tech/ynfy/config/thread/ThreadConstant.java
@@ -1,4 +1,4 @@
-package tech.ynfy.frame.constant;
+package tech.ynfy.config.thread;
 
 /**
  * 〈〉
